shop/requirement.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.0.6
django-jazzmin==2.5.0
django-modeltranslation==0.18.4
idna==3.3
Pillow==9.2.0
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
tzdata==2022.1
urllib3==1.26.12
